Initial commit
I like to start working on a project a bit before my first commit so the
basic file structure and idea is there. So here it is.

As of now I have implemented:
A system for storing data about planets
A tracker for resources
A tracker for owned bodies
A pretty pixel art solar system that is currently functionless

I have to decide which direction to go in: continue using html, or start
using <canvas>. And if I do use <canvas>, do I use html text on top of
that, or canvas text rendering?

Many decisions to go. We're a little over 15 hours in. But 7 of those
were sleep.