Cost 200 gas to attack; refactor purchase / build
There is a nice little tooltip that tells you that it costs, and it has
all the same messages. This is an important balance mechanic. The
refactor also separates purchasing from building, so it can be used for
this. This includes the "Insufficient..." message. Also, some of the
next patch is included for convenience, adding counters to a few
functions.