Make things more expensive; lots of $ for debug
So, first, I simply made things more expensive. This slows the game down
without changing balance at all. It makes it more fun in my opinion.

Second, I made the player start with a metric shit ton of everything in
the beginning, JUST FOR DEBUGGING. It makes it a lot easier; I'd been
going in and changing things on the console and it was bad.