Split `built` display into Contains and Defenses
Makes it easier to see military defenses at a glance / overall makes UI
better.

I didn't want to have to change the internals of `built` so
instead I split the displaying up using conditionals on the same list.
This made `drawList` virtually useless, and it definitely needs to be
refactored / reabstracted.