Educate player when attacking (fleet, mechanics)
It now has a tooltip when you mouse over attack that reminds you to
build a fleet.

If you attack and you don't have a fleet, instead of saying "Fleet
annihilated" it tells you how to build a fleet.

And last but not least, when attacking a big thing comes up that
explains the basics of attacking.