Allow AI to save up for attacking, increasing freq
team.nextBuild might be "attack" in which case it will save up to the
attackCost. This is handled by aiAttack anyway. This means that once a
team has a fleet, and they happen to select that, they will not buy
other things until they can attack. This will allow attacking to occur
at a more appropriate frequency.

As I was debugging this, I noticed that attacking with the new price is
now extremely uncommon. Thus, I was able to remove the attackChance and
fleetChance variables, further increasing frequency while not
overloading with way too much attacking.